+++ src/openpile/core/txt.py
@@ -19,11 +19,11 @@ def txt_pile(obj: Pile) -> str:
     txt = "{:-^80s}".format("")
     txt += "\n{:^80s}".format("Pile Input")
     txt += "\n{:-^80s}".format("")
-    txt += f"\nPile Material: {obj.material}"
+    txt += f"\nPile Material: {obj.material.name}"
     txt += f"\tPile Type: {obj.type}"
     txt += f"\tYoung modulus: {obj._young_modulus/1000:.0f} MPa"
-    txt += f"\tPoisson ratio = {obj._nu:.2f}"
-    txt += f"\nMaterial Unit Weight: {obj._uw:0.1f} kN/m3"
+    txt += f"\tPoisson ratio = {obj.material.poisson:.2f}"
+    txt += f"\nMaterial Unit Weight: {obj.material.unitweight:0.1f} kN/m3"
     txt += f"\n\nPile sections:\n"
     txt += obj.data.to_string(header=True, index=True)