+++ src/openpile/globals.py
@@ -1,2 +1,2 @@
 # version of the package
-VERSION = "0.7.1"
+VERSION = "0.8.0"